Add a multi-repository server
This merges a bunch of the ideas for a multi-repository CGI script and
simplifies things a bit.

It uses PATH_INFO so it generally wants paths of the form:

http://host/hg.cgi/virtualpath

This gets mapped via a simple config file to the real path. All info
about the repo is read from the repo's hgrc.