commands.revert: don't call hg.revert
commands.revert calculates everything that has to be done and then
calls hg.revert to checkout and remove files.  Unfortunately,
hg.revert has to recalculate everything and that can take a long
while, since it always operates on the whole working dir.

Changing commands.revert to manually checkout and remove files
makes things considerably faster, especially if we're reverting
a single file in a repo with a huge number of files.

This should be enough to close issue857.