win32text: be more careful about rejecting violating changesets
We now try to walk changesets in reverse order from newest to oldest,
so that if we see a file multiple times, we treat the newest version
as canonical.

This should prevent us from rejecting a changegroup that contains an
unacceptable commit followed later by a commit that fixes the problem.