help/templates: use hanging indent for field list
Now that the field list indentation is computed automatically, we no
longer need to align all fields the same. Using a simple hanging
indent makes it easy to edit the text in an editor like Emacs, which
will automatically recognized the hang and keep it.

The fields are still output with no blank lines between them.