tests: rework revision branch cache tests
The tests that were added with the revision branch cache in 678f53865c68 had
suffered from bit rot in the development iterations. They were no longer that
"good".

Now, the tests are rewritten and reworked to be more readable and maintainable
and relevant for the actual implementation.

This also utilizes the new 'f' helper tool for keeping an eye the cache files.