filemerge: split internal merge into api entry point and internal helper
This is a step toward adding 'union merge' to the internal merge tool.

'union merge' is a merge strategy which adds both left and right hand side
of a conflict region.  Git implements this merge strategy which is very
practical to have for merging to e.g. the Changelog file.