chgserver: auto exit after being idle for too long or lose the socket file
This is a part of the one server per config series. In multiple-server setup,
new server may be started for a temporary config change like in command line,
--config extensions.foo=bar.py. This may end up with a lot of not so useful
server processes. Other questions are about socket file and process management,
How to stop these processes? What if a new server wants to listen on a same
address, replacing the old one?

This patch introduces AutoExitMixIn, which will:

1. Exit after being idle for too long. So useless servers won't run forever.
2. Periodically check the ownership of socket file, exit if it is no longer
   owned. This brings strong consistency between the filesystem and the
   process, and handles some race conditions neatly.
   Since rename is atomic, a new server can just have a same server address
   with an old one and won't worry about how to make sure the old server is
   killed, address conflict, service downtime issues.
   The user can safely stop all servers by simply removing the socket files,
   without worrying about outdated or accidentally removed pidfiles.